CF 1.6
fixing float[] to double[] in ycbcr

+++ Catalano.Image/src/Catalano/Imaging/Concurrent/Filters/YCbCrFiltering.java
@@ -259,7 +259,7 @@ public class YCbCrFiltering implements IApplyInPlace{
                     int b = share.fastBitmap.getBlue(i, j);
                     
                     // convert to YCbCr
-                    float[] ycbcr = ColorConverter.RGBtoYCbCr(r, g, b, ColorConverter.YCbCrColorSpace.ITU_BT_601);
+                    double[] ycbcr = ColorConverter.RGBtoYCbCr(r, g, b, ColorConverter.YCbCrColorSpace.ITU_BT_601);
 
                     // check YCbCr values
                     if (
